Refactor ledger_zcash_crypto
