+++ ledger_zcash_crypto/Cargo.toml
@@ -7,8 +7,8 @@ edition = "2024"
 path = "src/lib.rs"
 
 [dependencies]
-hex = { version = "0.4.3", default-features =   false, features = ["alloc"] }
-ledger_device_sdk = { path = "../depz/ledger-device-rust-sdk/ledger_device_sdk", default-features = false, features = ["sys"] }
+#ledger_device_sdk = { version = "1.34.0", features = ["sys"] }
+ledger_device_sdk = { git = "https://github.com/LedgerHQ/ledger-device-rust-sdk.git", branch = "y333/zip32_derivation_support", features = ["sys"] }
 pasta_curves = { version = "0.5", default-features = false }
 
 [dev-dependencies]
